Workaround to compute metadata + file hashes
In python 2 unicode strings are represented with an explicit unicode
literal ('u'), whereas in python 3, although supported, it is dropped by
default. Why is this a problem? In python 2, the hash of (metadata + file
hash) would be computed using an input such as:
"[u'artist name', u'piece title', 'file hash']". In python 3 the same
data would be: "['artist name', 'piece title', 'file hash']". The fact
that the prefix 'u' is not there results in a different hash. In other
words, the inputs are different because the reprsentations are
different, which yields different outputs.