Use list comprehensions instead of filter
Quoting python docs:

"map() and filter() return iterators. If you really need a list and the
input sequences are all of equal length, a quick fix is to wrap map() in
list(), e.g. list(map(...)), but a better fix is often to use a list
comprehension (especially when the original code uses lambda), or
rewriting the code so it doesn’t need a list at all. [...]"

ref: https://docs.python.org/3/whatsnew/3.0.html#views-and-iterators-instead-of-lists